+++ src/main/resources/templates/users/details-edit.html
@@ -31,6 +31,10 @@
                             <input id="lName" name="lName" type="text" class="form-control" th:field="*{lName}"/>
                         </div>
                         <div class="form-group">
+                            <label for="birthdate" class="text-white">Birthdate</label>
+                            <input id="birthdate" name="birthdate" type="date" class="form-control" placeholder="Enter your Birthday" th:field="*{birthdate}" />
+                        </div>
+                        <div class="form-group">
                             <label for="city" class="text-white">City</label>
                             <input id="city" name="city" type="text" class="form-control" th:field="*{city}"/>
                         </div>
