+++ src/main/resources/templates/users/register.html
@@ -38,10 +38,6 @@
                             <span id="message"></span>
                         </div>
                         <div class="form-group">
-                            <label for="birthdate" class="text-white">Birthdate</label>
-                            <input id="birthdate" name="birthdate" type="date" class="form-control" placeholder="Enter your Birthday" th:field="*{birthdate}" />
-                        </div>
-                        <div class="form-group">
                             <label for="genderFormControlSelect" class="text-white">Gender</label>
                             <select class="form-control" id="genderFormControlSelect" th:field="*{gender}">
                                 <option value="disabled">Please Select</option>
