+++ src/main/resources/templates/users/show-user.html
@@ -24,13 +24,13 @@
                         <h1>User Information for: <span th:text="${user.username}"></span></h1>
                         <h2 th:text="${user.status}"></h2>
                         <h4 th:text="${'email: ' + user.email}"></h4>
-                        <h4 th:text="${'birthday: ' + user.birthdate}"></h4>
                         <h4 th:text="${'gender: ' + user.gender}"></h4>
 
                         <!-- =================== USER DETAIL INFO =================== -->
                         <th:block th:if="${user.userDetail != null}">
                             <h4 th:text="${user.userDetail.fName}"></h4>
                             <h4 th:text="${user.userDetail.lName}"></h4>
+                            <h4 th:text="${user.userDetail.birthdate}"></h4>
                             <h4 th:text="${user.userDetail.city}"></h4>
                             <h4 th:text="${user.userDetail.state}"></h4>
                             <h4 th:text="${user.userDetail.zipCode}"></h4>
@@ -48,15 +48,6 @@
                             <a th:href="'/users/details/create'" th:if="${showEditControls}">Add user details</a>
                         </th:block>
 
-                        <!-- Adding languages -->
-
-                        <form class="form-inline" th:action="'/users/show-user/add-language'" th:method="post">
-                            <div class="form-group mx-sm-3 mb-2">
-                                <input type="text" class="form-control" id="language" th:field="${languages.language}" placeholder="Add a language" />
-                            </div>
-                            <button type="submit" class="btn btn-outline-primary">Add</button>
-                        </form>
-
                         <!-- =================== USER AVAILABILITY INFO =================== -->
                         <th:block th:if="${user.availability != null}">
                             <h4 th:text="${user.availability.dayOfWeek}"></h4>
@@ -73,6 +64,17 @@
                             <h3>No User Availability</h3> <!-- Remove this -->
                             <a th:href="'/users/availability/create'" th:if="${showEditControls}">Add user Availability</a>
                         </th:block>
+
+                        <!-- Adding languages -->
+                        <th:block th:each="language : ${user.userLanguages}">
+                            <p th:text="${language.languages.language}"></p>
+                        </th:block>
+                        <form class="form-inline" th:action="'/users/show-user/add-language'" th:method="post">
+                            <div class="form-group mx-sm-3 mb-2">
+                                <input type="text" class="form-control" id="language" th:field="${languages.language}" placeholder="Add a language" />
+                            </div>
+                            <button type="submit" class="btn btn-outline-primary">Add</button>
+                        </form>
 
                     </div>
                 </div>
